Finished cleaning up 2021.

+++ src/main/kotlin/2021/DayTwentyfour.kt
@@ -60,7 +60,7 @@ class DayTwentyfour(file: String) : Project {
             output += validLastIndexes[i]!!.toString()
         }
 
-        return output.toInt()
+        return output.toLong()
 
     }
 }
